requirement.txt
PyPDF2>=3.0.1
Pillow>=10.0.0
pytesseract>=0.3.10
pandas>=2.0.0
openpyxl>=3.1.0
faiss-cpu>=1.7.4
numpy>=1.24.0
oracledb>=1.4.0
pyyaml>=6.0.0
jsonschema>=4.17.0
tenacity>=8.2.0
pytest>=7.4.0
azure-ai-inference>=1.0.0
azure-core>=1.26.0
google-cloud-aiplatform>=1.35.0